# lib/api-zod/src/generated/types/setWorkLocationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SetWorkLocationBody {
  name: string;
  latitude: number;
  longitude: number;
  radiusMeters: number;
}