# lib/api-zod/src/generated/types/timeEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TimeEntry {
  id: number;
  clockIn: Date;
  clockOut: Date | null;
  durationMinutes: number | null;
  notes: string | null;
  latitude: number | null;
  longitude: number | null;
  createdAt: Date;
}