# lib/api-zod/src/generated/types/totals.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Totals {
  totalMinutes: number;
  totalHours: number;
  totalDays: number;
  totalEntries: number;
  averageHoursPerDay: number;
}