# lib/api-zod/src/generated/types/updateTimeEntryBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateTimeEntryBody {
  clockOut?: Date;
  notes?: string;
  latitude?: number;
  longitude?: number;
}