# lib/api-zod/src/generated/types/weeklySummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { WeeklySummaryDailyBreakdownItem } from "./weeklySummaryDailyBreakdownItem";

export interface WeeklySummary {
  weekStart: Date;
  weekEnd: Date;
  totalMinutes: number;
  totalHours: number;
  daysWorked: number;
  dailyBreakdown: WeeklySummaryDailyBreakdownItem[];
}